添加微信其他模块

+++ Inject/Communication/Packet/XXFilePacket.m
@@ -61,7 +61,7 @@
     }
     infoDic[@"XXFileOffset"] = [NSNumber numberWithLongLong:didReadLenght];
 
-    NSData * readData = [fileHandle readDataOfLength:readLenght];
+    NSData * readData = [fileHandle readDataOfLength:(NSUInteger)readLenght];
     *data = readData;
     
     infoDic[@"XXFileDataLenght"] = [NSNumber numberWithLongLong:readData.length];
